Create a temporary directory with File::Temp
t/21server-list.t fails if run from a read-only location. The reason
is that it attempts to create a temporary directory in a current
working directory.

This patch fixes it by using File::Temp. That core Perl module
utilizes a location dedicated for temporary files which is usually
writable. That module also simplifies cleaning the directory.